chore(shared): migrate auth types to shared package
