+++ audit-report.md
@@ -776,16 +776,16 @@ The database uses PostgreSQL with Drizzle ORM for schema definition and query bu
 ## 1. Type System Unification (High Priority)
 
 ### 1.1 Shared Types Migration (2-3 weeks)
-- [ ] Create a dedicated `@shared/types` package
+- [x] Create a dedicated `@shared/types` package
 - [ ] Migrate all type definitions to the shared package
-- [ ] Set up proper versioning for the shared package
-- [ ] Update both client and server to consume the shared types
+- [x] Set up proper versioning for the shared package
+- [x] Update both client and server to consume the shared types
 
 ### 1.2 Type Safety Enhancements (1-2 weeks)
-- [ ] Enable TypeScript's `strict` mode
+- [x] Enable TypeScript's `strict` mode
 - [ ] Add ESLint rules to prevent `any` usage
 - [ ] Implement runtime type validation with Zod
-- [ ] Add type checking to CI/CD pipeline
+- [x] Add type checking to CI/CD pipeline
 
 ## 2. Authentication & Authorization (High Priority)
 
@@ -804,7 +804,7 @@ The database uses PostgreSQL with Drizzle ORM for schema definition and query bu
 ## 3. API Contract Consistency (Medium Priority)
 
 ### 3.1 Request/Response Validation (2 weeks)
-- [ ] Implement request validation with Zod
+- [x] Implement request validation with Zod
 - [ ] Add response validation
 - [ ] Generate OpenAPI/Swagger documentation
 - [ ] Add contract testing
@@ -849,7 +849,7 @@ The database uses PostgreSQL with Drizzle ORM for schema definition and query bu
 
 ### 7.1 API Performance (1-2 weeks)
 - [ ] Implement response caching
-- [ ] Add request/response compression
+- [x] Add request/response compression
 - [ ] Optimize database queries
 - [ ] Implement pagination
 
@@ -870,7 +870,7 @@ The database uses PostgreSQL with Drizzle ORM for schema definition and query bu
 ## 9. Deployment & Monitoring (High Priority)
 
 ### 9.1 CI/CD Pipeline (1-2 weeks)
-- [ ] Set up automated testing
+- [x] Set up automated testing
 - [ ] Implement blue-green deployments
 - [ ] Add rollback capabilities
 - [ ] Monitor deployment health
