+++ shared/types/auth/dto/index.ts
@@ -2,4 +2,5 @@
 export * from './login.dto.js';
 export * from './register.dto.js';
 export * from './password-reset.dto.js';
+export * from './user-response.dto.js';
 // Add other DTO exports here
