+++ shared/types/auth/dto/user-response.dto.ts
@@ -1,0 +1,10 @@
+export interface UserResponse {
+  id: string;
+  email: string;
+  role: import('../permissions.js').UserRole;
+  firstName: string | null;
+  lastName: string | null;
+  emailVerified: boolean;
+  createdAt: string;
+  updatedAt: string;
+}
