+++ shared/types/index.ts
@@ -3,8 +3,10 @@ export { AuthError } from './auth/auth.js';
 export type { AuthErrorCode, AuthState, AuthConfig } from './auth/auth.js';
 export type { AuthTokens, JwtPayload, TokenType } from './auth/jwt.js';
 export type { UserRole } from './auth/permissions.js';
+export type { AuthResponse } from "./auth/jwt.js";
 export type { AuthUser, UserProfile, UserSettings } from './auth/user.js';
 // Export DTOs
 export type { LoginDto } from './auth/dto/login.dto.js';
 export type { RegisterDto } from './auth/dto/register.dto.js';
 export type { RequestPasswordResetDto, ResetPasswordDto, ChangePasswordDto } from './auth/dto/password-reset.dto.js';
+export type { UserResponse } from "./auth/dto/user-response.dto.js";
